Sumando dos mantenedores mas

+++ app/Http/Controllers/UniversidadController.php
@@ -9,61 +9,44 @@ use Illuminate\Support\Facades\Storage;
 
 class UniversidadController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
         $universidades = Universidad::all();
-        $comunas = Comuna::orderBy('nombre_comuna', 'asc')->get();
+        $comunas = Comuna::all();
         return view('layouts.university.university_index', compact('universidades', 'comunas'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
-        try {
-            $validatedData = $request->validate([
-                'nombre_universidad' => 'required|string|max:255',
-                'direccion_universidad' => 'required|string|max:255',
-                'telefono_universidad' => 'required|string|max:15',
-                'imagen_logo' => 'nullable|image|mimes:jpg,jpeg,png,gif|max:2048',
-                'id_comuna' => 'required|exists:comunas,id',
-            ]);
+        $validatedData = $request->validate([
+            'id_universidad' => 'required|string|max:255',
+            'nombre_universidad' => 'required|string|max:255',
+            'direccion_universidad' => 'required|string|max:255',
+            'telefono_universidad' => 'required|regex:/^[0-9+]+$/|max:15',
+            'imagen_logo' => 'nullable|image|mimes:jpg,jpeg,png,gif|max:2048',
+            'comunas_id' => 'required|exists:comunas,id',
+        ]);
 
-            $imagenPath = null;
-            if ($request->hasFile('imagen_logo')) {
-                $imagenPath = $request->file('imagen_logo')->store('logos', 'public');
-            }
+        $imagenPath = null;
 
-            Universidad::create([
-                'nombre_universidad' => $validatedData['nombre_universidad'],
-                'direccion_universidad' => $validatedData['direccion_universidad'],
-                'telefono_universidad' => $validatedData['telefono_universidad'],
-                'imagen_logo' => $imagenPath,
-                'id_comuna' => $validatedData['id_comuna'],
-            ]);
-
-            return redirect()->route('universidades.index')->with('success', 'Universidad creada exitosamente.');
-
-        } catch (\Exception $e) {
-            return redirect()->back()->withErrors(['error' => 'Hubo un problema al crear la universidad. Por favor, intente nuevamente.'])->withInput();
+        if ($request->hasFile('imagen_logo')) {
+            $file = $request->file('imagen_logo');
+            $nombreArchivo = str_replace(' ', '_', strtolower($validatedData['nombre_universidad'])) . '.' . $file->getClientOriginalExtension();
+            $imagenPath = $file->storeAs('imagenes/logo_universidad', $nombreArchivo, 'public');
         }
+
+        Universidad::create([
+            'id_universidad' => $validatedData['id_universidad'],
+            'nombre_universidad' => $validatedData['nombre_universidad'],
+            'direccion_universidad' => $validatedData['direccion_universidad'],
+            'telefono_universidad' => $validatedData['telefono_universidad'],
+            'imagen_logo' => $imagenPath,
+            'comunas_id' => $validatedData['comunas_id'],
+        ]);
+
+        return redirect()->route('universities.index')->with('success', 'Universidad creada exitosamente.');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit($id)
     {
         $universidad = Universidad::findOrFail($id);
@@ -71,61 +54,53 @@ class UniversidadController extends Controller
         return view('layouts.university.university_edit', compact('universidad', 'comunas'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, $id)
-{
-    try {
+    {
         $validatedData = $request->validate([
+            'id_universidad' => 'required|string|max:255',
             'nombre_universidad' => 'required|string|max:255',
             'direccion_universidad' => 'required|string|max:255',
-            'telefono_universidad' => 'required|string|max:15',
+            'telefono_universidad' => 'required|regex:/^[0-9+]+$/|max:15',
             'imagen_logo' => 'nullable|image|mimes:jpg,jpeg,png,gif|max:2048',
-            'id_comuna' => 'required|exists:comunas,id',
+            'comunas_id' => 'required|exists:comunas,id',
         ]);
 
         $universidad = Universidad::findOrFail($id);
+        $imagenPath = $universidad->imagen_logo;
 
         if ($request->hasFile('imagen_logo')) {
-            // Elimina la imagen antigua si existe
-            if ($universidad->imagen_logo) {
+            if ($universidad->imagen_logo && Storage::disk('public')->exists($universidad->imagen_logo)) {
                 Storage::disk('public')->delete($universidad->imagen_logo);
             }
-            $imagenPath = $request->file('imagen_logo')->store('logos', 'public');
-        } else {
-            $imagenPath = $universidad->imagen_logo;
+
+            // Guardar la nueva imagen
+            $file = $request->file('imagen_logo');
+            $nombreArchivo = str_replace(' ', '_', strtolower($validatedData['nombre_universidad'])) . '.' . $file->getClientOriginalExtension();
+            $imagenPath = $file->storeAs('imagenes/logo_universidad', $nombreArchivo, 'public');
         }
 
         $universidad->update([
+            'id_universidad' => $validatedData['id_universidad'],
             'nombre_universidad' => $validatedData['nombre_universidad'],
             'direccion_universidad' => $validatedData['direccion_universidad'],
             'telefono_universidad' => $validatedData['telefono_universidad'],
             'imagen_logo' => $imagenPath,
-            'id_comuna' => $validatedData['id_comuna'],
+            'comunas_id' => $validatedData['comunas_id'],
         ]);
 
-        return redirect()->route('universidades.index')->with('success', 'Universidad actualizada exitosamente.');
-
-    } catch (\Exception $e) {
-        return redirect()->back()->withErrors(['error' => 'Hubo un problema al actualizar la universidad. Por favor, intente nuevamente.'])->withInput();
+        return redirect()->route('universities.index')->with('success', 'Universidad actualizada exitosamente.');
     }
-}
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy($id)
     {
-        try {
-            $universidad = Universidad::findOrFail($id);
-            $universidad->delete();
-            return redirect()->route('universitys.index')->with('success', 'Universidad eliminado exitosamente.');
+        $universidad = Universidad::findOrFail($id);
 
-        } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
-            return response()->json(['success' => false, 'message' => 'Universidad no encontrado.'], 404);
-        } catch (\Exception $e) {
-            return response()->json(['success' => false, 'message' => 'Error al borrar la Universidad: ' . $e->getMessage()], 500);
+        if ($universidad->imagen_logo && Storage::disk('public')->exists($universidad->imagen_logo)) {
+            Storage::disk('public')->delete($universidad->imagen_logo);
         }
+
+        $universidad->delete();
+
+        return redirect()->route('universities.index')->with('success', 'Universidad eliminada exitosamente.');
     }
 }
